refactor(auth): dedupe unique-violation helper and share LDAP sign-in wire type
Extract the byte-identical isUniqueViolation/MaybeCodedError helper from
ldap-provisioning.ts and ldap-reconciliation.ts into src/db/unique-violation.ts.
Move the LDAP sign-in success response shape into @hashhive/shared as
ldapSignInSuccessSchema (roles reusing the UserRole union) per the
wire-shapes-live-in-shared convention, replacing the hand-declared,
stringly-typed frontend interface. Behavior-preserving.
